refactor(Box): use Box to replace WatchedBox and ResultsBox
